requirements.txt
requests==2.32.3
beautifulsoup4==4.11.2
configparser==5.3.0
Flask==3.0.3
line-bot-sdk==3.17.0
waitress==2.1.2
pymysql==1.0.3
waitress==2.1.2
openai==0.27.8